server/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
pyzmq>=22.0.0
numpy>=1.21.0
